hoist handleSearchSelect into context
