+++ src/ManageInventory/ManageInventory.js
@@ -1,11 +1,7 @@
 import React, { useContext } from 'react';
 import SearchBar from '../common/SearchBar';
-import axios from 'axios';
-import makeAuthHeader from '../utils/makeAuthHeader';
 import ManageInventoryList from './ManageInventoryList';
 import { Segment, Header, Icon, Divider } from 'semantic-ui-react';
-import { GET_CARDS_WITH_INFO } from '../utils/api_resources';
-import { InventoryCard } from '../utils/ScryfallCard';
 import styled from 'styled-components';
 import AllLocationInventory from './AllLocationInventory';
 import { InventoryContext } from '../context/InventoryContext';
@@ -16,22 +12,7 @@ const HeaderContainer = styled('div')({
 });
 
 export default function ManageInventory() {
-    const { searchResults, setSearchResults } = useContext(InventoryContext);
-
-    const handleSearchSelect = async (term) => {
-        try {
-            const { data } = await axios.get(GET_CARDS_WITH_INFO, {
-                params: { title: term, matchInStock: false },
-                headers: makeAuthHeader(),
-            });
-
-            const modeledData = data.map((c) => new InventoryCard(c));
-
-            setSearchResults(modeledData);
-        } catch (e) {
-            console.log(e);
-        }
-    };
+    const { searchResults, handleSearchSelect } = useContext(InventoryContext);
 
     return (
         <>
