+++ src/context/InventoryContext.js
@@ -1,9 +1,28 @@
+import axios from 'axios';
 import React, { createContext, useState } from 'react';
+import makeAuthHeader from '../utils/makeAuthHeader';
+import { InventoryCard } from '../utils/ScryfallCard';
+import { GET_CARDS_WITH_INFO } from '../utils/api_resources';
 
 export const InventoryContext = createContext();
 
 export function InventoryProvider({ children }) {
     const [searchResults, setSearchResults] = useState([]);
+
+    const handleSearchSelect = async (term) => {
+        try {
+            const { data } = await axios.get(GET_CARDS_WITH_INFO, {
+                params: { title: term, matchInStock: false },
+                headers: makeAuthHeader(),
+            });
+
+            const modeledData = data.map((c) => new InventoryCard(c));
+
+            setSearchResults(modeledData);
+        } catch (e) {
+            console.log(e);
+        }
+    };
 
     const changeCardQuantity = (id, qoh) => {
         const copiedState = [...searchResults];
@@ -14,7 +33,7 @@ export function InventoryProvider({ children }) {
 
     return (
         <InventoryContext.Provider
-            value={{ searchResults, setSearchResults, changeCardQuantity }}
+            value={{ searchResults, changeCardQuantity, handleSearchSelect }}
         >
             {children}
         </InventoryContext.Provider>
